Refactor tests/
